requirements.txt
requests
python-dotenv
flask
Flask-Cors
rich